Move card to folder
