+++ src/Card/index.jsx
@@ -1,32 +1,6 @@
-import './reset.css';
-import './App.css';
-import React, { useState, useEffect, useRef } from 'react';
-
-import './Card.css';
-
-// https://github.com/heyitsarpit/react-hooks-library/blob/main/packages/core/useHover/index.ts
-const useHover = (ref) => {
-  const [isHovered, setIsHovered] = useState(false);
-
-  useEffect(() => {
-    const el = ref.current;
-
-    if (!el) return () => {};
-
-    const onMouseEnter = () => setIsHovered(true);
-    const onMouseLeave = () => setIsHovered(false);
-
-    el.addEventListener('mouseenter', onMouseEnter);
-    el.addEventListener('mouseleave', onMouseLeave);
-
-    return () => {
-      el.removeEventListener('mouseenter', onMouseEnter);
-      el.removeEventListener('mouseleave', onMouseLeave);
-    };
-  }, [ref]);
-
-  return isHovered;
-};
+import React, { useEffect, useRef } from 'react';
+import useHover from '../hooks/useHover';
+import './styles.css';
 
 export default function Card(props) {
   const { card, equipped, ...rest } = props;
