+++ src/Card/styles.css
@@ -110,4 +110,4 @@
   background-repeat: no-repeat;
   pointer-events: none;
   opacity: 0.8;
-}
+}
